requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
pydantic>=2.12.5
jinja2>=3.1.6
python-multipart>=0.0.21
pillow>=12.1.0
requests>=2.32.4